get_next_line mandatory and bonus
This get_next_line passes all test, including fracinette strict.
The timeouts in the strict mode can be check at the francinette
repository fac, it's a well-known problem for this test.
Added an static array of 256 file descriptors, the numbers is choosen
using the command ulimit -n, which prints the maximum numbers of fd
allowed to be open an a certain time.